siggen: Fix dump of non-runtime sigdata
If the debug code is uncommented in SignatureGeneratorBasic.finalise(),
sigtask stamp file are created without runtime data such as
'runtaskdeps' and 'file_checksum_values'. If we then try to run
bitbake-dumpsig on one of these files, we will see a traceback as the
expected keys are not present in the sigdata dictionary:

    Traceback (most recent call last):
      File ".../bitbake/bin/bitbake-dumpsig", line 195, in <module>
        output = bb.siggen.dump_sigfile(options.sigdatafile1)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File ".../bitbake/lib/bb/siggen.py", line 1266, in dump_sigfile
        computed_taskhash = calc_taskhash(a_data)
                            ^^^^^^^^^^^^^^^^^^^^^
      File ".../bitbake/lib/bb/siggen.py", line 1199, in calc_taskhash
        for dep in sigdata['runtaskdeps']:
                   ~~~~~~~^^^^^^^^^^^^^^^
    KeyError: 'runtaskdeps'

To fix this, do not try to compute a task hash if the required data is
not present. We can key off 'runtaskdeps' as this will likely always be
needed to calculate a valid task hash.

Signed-off-by: Paul Barker <paul@pbarker.dev>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>