toaster: remove dependency on poky distro
The poky distro is now stripped of its repositories. Toaster had used that
repo as both a quick way to get the three core layers (openembedded-code,
meta-poky, and meta-yocto-bsp) but also a way to get the related bitbake,
which was important because bitbake is not in the Layer Index. Now, Toaster
must explicitly clone each of the core layers plus the proper bitbake layer.

Notes on "orm/fixtures":
  * The "orm/fixtures" files used to hard code the three core layers to
    point to the poky distro, ignoring the Layer Index values
  * The script "lsupdates.py" is now allowed to use the Layer Index to
    properly initialize these layers.
  * The fixture files are still used because (a) they define the supported
    releases for Toaster, plus (b) they now define the proper bitbake repo
    path and branches to associate with the YP releases.
Notes on "localhostbecontroller.py":
  * Always insert the proper bitbake into the layer clone list
  * Insure that both 'BBBASEDIR' and 'BITBAKEDIR' are defined in the
    sub-shells so that oe-init finds the right bitbake (e.g.
    'oe-buildenv-internal')

[YOCTO #16012]

Signed-off-by: David Reyna <David.Reyna@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>