doc/bitbake-setup: show how to clone bitbake before using bitbake-setup
Users reading the documentation on the website or else might not have a
copy of bitbake yet. It is not straightforward to users that
bitbake-setup is part of the source tree and that it is the only
location to use it from (for now).

Add a step at the beginning of the Quick Start section that shows how to
clone bitbake-setup. The default branch (master) will be cloned, which
is currently what is recommended. This might be subject to change in the
future if bitbake-setup becomes backwards-incompatible.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>