bitbake-setup: Allow setup-dir-name to be specified in JSON (and use that in the default registry)
Enables the JSON file to specify the preferred name of the setup
directory with a 'setup-dir-name' key. This key can have variable
expansions in the same format as python string.Template strings (which
matches the shell variable expansion rules).

Variables that can be expanded are any fragment configuration prompted
by the user (e.g. the keys in "oe-fragments-one-of"), or "$name" to use
the name of the configuration (e.g. the .conf.json file).

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

default-registry: Change default setup dir

Uses the template mechanism to provide shorter setup-dir names by
default. Specifically, the machine names are no longer part of the setup
path since they are intended to be changed without needing a new setup
directory.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

Further tweaks by Alex:
- $name support is not actually implemented (as far as I see)
- add fallbacks to the full name if the short name already exists in the filesystem
or there's a setting that forces full names
- add documentation for the setup dir entry in configs and the setting to ignore it
- adjust the schema
- adjust the selftest to include the feature and test for it

RP: Simplify the poky default string to be $distro-master

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>