bitbake-setup: use a transient temporary top_dir to list available configs
As pointed out here:
https://lists.openembedded.org/g/bitbake-devel/topic/116374639

reusing the standard top_dir is prone to issues: the user, if they're just starting,
may not want anything to be left over after issuing 'list' or even want to decide where
the top_dir should be, they just want a list of configs.

Also if the top_dir is set to . (which is the default) and the user is in
a directory that cannot be written into, there'll be a baffling traceback.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>