bitbake-setup: correct the confusion over 'uri' format in json configs
The 'uri' property in bitbake-setup json configurations is actually
using the standard git URI syntax, and not bitbake fetcher-specific one.
Bitbake-setup converts it to something the git fetcher understands.

This was overlooked when default registry configs were added. It worked
but only by coincidence (and was using git:// protocol despite 'protocol=https'
in the value).

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>