bitbake-setup: add an optional 'expires' property to the schema
This can be used to document, in a progammatic way, the end of life
date for a configuration, and make users aware of it.

Bitbake documentation already has its description and does not need
an update.

Bitbake-setup is using the property to tell the users until when
a given configuration is supported, and omits it from available
interactive choices if the date has passed (unless asked expicitly
on command line).

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>