bitbake-setup: convert print calls to logger
We already have a logger defined but don't use it for logging. Convert
existing print calls to use the logger, categorizing them by importance
with the levels we have.

Most of the calls are replaced with logger.plain, which renders the same
way as print. Other messages are preceded by their level (NOTE:,
WARNING:, etc.).

This does not change the verbosity of bitbake-setup, all current
messages are still printed.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>